31.03.2022

+++ ft_bzero.c
@@ -6,21 +6,22 @@
 /*   By: fyuzhyk <fyuzhyk@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/28 14:58:10 by fyuzhyk           #+#    #+#             */
-/*   Updated: 2022/03/29 10:47:53 by fyuzhyk          ###   ########.fr       */
+/*   Updated: 2022/03/31 10:44:50 by fyuzhyk          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void ft_bzero(void *s, unsigned int n)
+#include "libft.h"
+void	ft_bzero(void *s, size_t n)
 {
-    unsigned int i;
+	unsigned int	i;
 
-    i = 0;
-    while (i < n)
-    {
-        *(unsigned char *)s = '\0';
-        s++;
-        i++;
-    }
+	i = 0;
+	while (i < n)
+	{
+		*(unsigned char *)s = '\0';
+		s++;
+		i++;
+	}
 }
 
 // #include <stdio.h>
