+++ ft_memcpy.c
@@ -6,11 +6,12 @@
 /*   By: fyuzhyk <fyuzhyk@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/28 14:59:07 by fyuzhyk           #+#    #+#             */
-/*   Updated: 2022/03/29 11:01:13 by fyuzhyk          ###   ########.fr       */
+/*   Updated: 2022/03/31 10:24:04 by fyuzhyk          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void *ft_memcpy(void *dest, const void *src, unsigned int n)
+#include "libft.h"
+void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
 	unsigned int	i;
